Improve scrolling behaviour on iOS (and possibly Chrome)
SharePoint makes #s4-workspace the window height and scrollable in order to keep the suite bar and ribbon at the top of the window at all times. This wastes limited screen space on the phone in two ways:
1. The space used by the suite bar and ribbon is always occupied.
2. It can't take advantage of the fact that iOS will shrink the window header and remove the toolbar as the page is scrolled - but only if the element being scrolled is the body.

Setting the #s4-workspace height to auto allows the body to become scrollable and providing more screen space for content.

(I suspect Chrome does a similar thing but don't have access to a device to test)